+++ Neolution.DotNet.Console.UnitTests/Common/Fakes/DefaultOptions.cs
@@ -1,0 +1,12 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.Common.Fakes
+{
+    using CommandLine;
+
+    /// <summary>
+    /// The options stub for the <see cref="DefaultCommand"/>
+    /// </summary>
+    [Verb("default", isDefault: true)]
+    public class DefaultOptions
+    {
+    }
+}