+++ Neolution.DotNet.Console.UnitTests/Common/Stubs/ScopedServiceStub.cs
@@ -1,14 +1,17 @@
 ﻿namespace Neolution.DotNet.Console.UnitTests.Common.Stubs
 {
+    using System.Threading.Tasks;
+
     /// <summary>
     /// Stub for a generic scoped service implementation.
     /// </summary>
     public class ScopedServiceStub : IScopedServiceStub
     {
         /// <inheritdoc />
-        public void DoSomething()
+        public async Task DoSomethingAsync()
         {
             // do nothing
+            await Task.CompletedTask;
         }
     }
 }