+++ Neolution.DotNet.Console.UnitTests/RunTests/Stubs/InjectServicesWithVariousLifetimesOptionsStub.cs
@@ -1,0 +1,13 @@
+﻿namespace Neolution.DotNet.Console.UnitTests.RunTests.Stubs
+{
+    using CommandLine;
+    using Neolution.DotNet.Console.UnitTests.Common;
+
+    /// <summary>
+    /// The options stub for the <see cref="InjectServicesWithVariousLifetimesCommandStub"/>
+    /// </summary>
+    [Verb(UnitTestsConstants.InjectServicesWithVariousLifetimes)]
+    public class InjectServicesWithVariousLifetimesOptionsStub
+    {
+    }
+}